ai-backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
google-generativeai==0.8.5
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-ai-generativelanguage==0.6.15
grpcio==1.73.1
grpcio-status==1.71.2
requests==2.32.4
tqdm==4.67.1